Fixed the date configuration for fetching the data
